requirements.txt

keras==2.12.0
numpy==1.23.5
pandas==2.0.1
Pillow==9.5.0
streamlit==1.22.0
tensorflow==2.12.0
uvicorn==0.22.0

